requirements.txt
websocket-client
streamlit